+++ src/backend/kernel_agents/agent_base.py
@@ -1,14 +1,15 @@
 import logging
 import json
 import os
-from typing import Any, Dict, List, Mapping, Optional, Callable, Awaitable
+from typing import Any, Dict, List, Mapping, Optional, Callable, Awaitable, Union
 
 import semantic_kernel as sk
 from semantic_kernel.functions import KernelFunction
 from semantic_kernel.kernel_arguments import KernelArguments
 # Import core components needed for Semantic Kernel plugins
 from semantic_kernel.plugin_definition import kernel_function, kernel_function_context_parameter
-
+# Import Azure AI Agent
+from semantic_kernel.agents.azure_ai.azure_ai_agent import AzureAIAgent
 
 # Import Pydantic model base
 from semantic_kernel.kernel_pydantic import KernelBaseModel
@@ -21,13 +22,14 @@ from models.messages_kernel import (
     Step,
     StepStatus,
 )
+from config_kernel import Config
 from event_utils import track_event_if_configured
 
 # Default formatting instructions used across agents
 DEFAULT_FORMATTING_INSTRUCTIONS = "Instructions: returning the output of this function call verbatim to the user in markdown. Then write AGENT SUMMARY: and then include a summary of what you did."
 
-class BaseAgent(KernelBaseModel):
-    """BaseAgent implemented using Semantic Kernel instead of AutoGen."""
+class BaseAgent:
+    """BaseAgent implemented using Semantic Kernel with Azure AI Agent support."""
 
     def __init__(
         self,
@@ -39,13 +41,27 @@ class BaseAgent(KernelBaseModel):
         tools: Optional[List[KernelFunction]] = None,
         system_message: Optional[str] = None,
         agent_type: Optional[str] = None,
+        use_azure_ai_agent: bool = True,
     ):
-        super().__init__()
+        """Initialize the base agent.
+        
+        Args:
+            agent_name: The name of the agent
+            kernel: The semantic kernel instance
+            session_id: The session ID
+            user_id: The user ID
+            memory_store: The memory context for storing agent state
+            tools: Optional list of tools for the agent
+            system_message: Optional system message for the agent
+            agent_type: Optional agent type string for automatic tool loading
+            use_azure_ai_agent: Whether to use Azure AI Agent or regular function calling
+        """
         self._agent_name = agent_name
         self._kernel = kernel
         self._session_id = session_id
         self._user_id = user_id
         self._memory_store = memory_store
+        self._use_azure_ai_agent = use_azure_ai_agent
         
         # If agent_type is provided, load tools from config automatically
         if agent_type and not tools:
@@ -61,9 +77,22 @@ class BaseAgent(KernelBaseModel):
         self._system_message = system_message or self._default_system_message()
         self._chat_history = [{"role": "system", "content": self._system_message}]
         
-        # Log initialization
-        logging.info(f"Initialized {agent_name} with {len(self._tools)} tools")
+        # If using Azure AI Agent, initialize it
+        if self._use_azure_ai_agent:
+            self._agent = Config.CreateAzureAIAgent(
+                kernel=self._kernel,
+                agent_name=self._agent_name,
+                instructions=self._system_message
+            )
+            
+            # Register tools with the agent
+            for tool in self._tools:
+                self._agent.add_function(tool)
         
+        # Log initialization
+        logging.info(f"Initialized {agent_name} with {len(self._tools)} tools, using {'Azure AI Agent' if use_azure_ai_agent else 'standard function calling'}")
+        
+        # Register the handler functions
         self._register_functions()
 
     def _default_system_message(self) -> str:
@@ -72,8 +101,12 @@ class BaseAgent(KernelBaseModel):
 
     def _register_functions(self):
         """Register this agent's functions with the kernel."""
-        # Register the action handler as a native function
-        self._kernel.import_skill(self, skill_name=self._agent_name)
+        # Register the action handler as a kernel function
+        self._kernel.add_function(
+            self.handle_action_request, 
+            plugin_name=self._agent_name, 
+            function_name="handle_action_request"
+        )
 
     @staticmethod
     def create_dynamic_function(name: str, response_template: str, formatting_instr: str = DEFAULT_FORMATTING_INSTRUCTIONS) -> Callable[..., Awaitable[str]]:
@@ -112,7 +145,7 @@ class BaseAgent(KernelBaseModel):
             A dictionary containing the configuration
         """
         if config_path is None:
-            # Default path relative to the caller's file
+            # Default path relative to the tools directory
             current_dir = os.path.dirname(os.path.abspath(__file__))
             backend_dir = os.path.dirname(os.path.dirname(current_dir))
             config_path = os.path.join(backend_dir, "tools", f"{agent_type}_tools.json")
@@ -121,7 +154,7 @@ class BaseAgent(KernelBaseModel):
             with open(config_path, "r") as f:
                 return json.load(f)
         except Exception as e:
-            print(f"Error loading {agent_type} tools configuration: {e}")
+            logging.error(f"Error loading {agent_type} tools configuration: {e}")
             # Return empty default configuration
             return {
                 "agent_name": f"{agent_type.capitalize()}Agent",
@@ -163,20 +196,11 @@ class BaseAgent(KernelBaseModel):
         
         return kernel_functions
 
-    @kernel_function(
-        description="Handle an action request from another agent",
-        name="handle_action_request",
-    )
-    @kernel_function_context_parameter(
-        name="action_request_json",
-        description="JSON string of the action request",
-    )
     async def handle_action_request(
-        self, context: KernelArguments
+        self, action_request_json: str
     ) -> str:
         """Handle an action request from another agent or the system."""
         try:
-            action_request_json = context["action_request_json"]
             action_request = ActionRequest.parse_raw(action_request_json)
             
             step: Optional[Step] = await self._memory_store.get_step(
@@ -191,27 +215,20 @@ class BaseAgent(KernelBaseModel):
                 )
                 return response.json()
             
-            self._chat_history.extend([
-                {"role": "assistant", "content": action_request.action, "name": "GroupChatManager"},
-                {"role": "user", "content": f"{step.human_feedback}. Now make the function call", "name": "HumanAgent"},
-            ])
-            
             try:
-                variables = KernelArguments()
-                variables["step_id"] = action_request.step_id
-                variables["session_id"] = action_request.session_id
-                variables["plan_id"] = action_request.plan_id
-                variables["action"] = action_request.action
-                variables["chat_history"] = str(self._chat_history)
-                
-                result = await self._execute_tool_with_llm(variables)
-                
+                # Choose execution method based on configuration
+                if self._use_azure_ai_agent:
+                    result_content = await self._execute_with_azure_ai_agent(step, action_request)
+                else:
+                    result_content = await self._execute_with_function_calling(step, action_request)
+
+                # Store agent message in cosmos
                 await self._memory_store.add_item(
                     AgentMessage(
                         session_id=action_request.session_id,
                         user_id=self._user_id,
                         plan_id=action_request.plan_id,
-                        content=f"{result}",
+                        content=f"{result_content}",
                         source=self._agent_name,
                         step_id=action_request.step_id,
                     )
@@ -223,14 +240,15 @@ class BaseAgent(KernelBaseModel):
                         "session_id": action_request.session_id,
                         "user_id": self._user_id,
                         "plan_id": action_request.plan_id,
-                        "content": f"{result}",
+                        "content": f"{result_content}",
                         "source": self._agent_name,
                         "step_id": action_request.step_id,
                     },
                 )
                 
+                # Update the step
                 step.status = StepStatus.completed
-                step.agent_reply = result
+                step.agent_reply = result_content
                 await self._memory_store.update_step(step)
                 
                 track_event_if_configured(
@@ -238,31 +256,33 @@ class BaseAgent(KernelBaseModel):
                     {
                         "status": StepStatus.completed,
                         "session_id": action_request.session_id,
-                        "agent_reply": f"{result}",
+                        "agent_reply": f"{result_content}",
                         "user_id": self._user_id,
                         "plan_id": action_request.plan_id,
-                        "content": f"{result}",
+                        "content": f"{result_content}",
                         "source": self._agent_name,
                         "step_id": action_request.step_id,
                     },
                 )
                 
+                # Create the response
                 action_response = ActionResponse(
                     step_id=step.id,
                     plan_id=step.plan_id,
                     session_id=action_request.session_id,
-                    result=result,
+                    result=result_content,
                     status=StepStatus.completed,
                 )
                 
+                # Publish to group chat manager
                 await self._publish_to_group_chat_manager(action_response)
                 
                 return action_response.json()
                 
             except Exception as e:
-                logging.exception(f"Error during tool execution: {e}")
+                logging.exception(f"Error during agent execution: {e}")
                 track_event_if_configured(
-                    "Base agent - Error during tool execution, captured into the cosmos",
+                    "Base agent - Error during execution, captured into the cosmos",
                     {
                         "session_id": action_request.session_id,
                         "user_id": self._user_id,
@@ -290,8 +310,72 @@ class BaseAgent(KernelBaseModel):
                 message=f"Error handling action request: {str(e)}"
             ).json()
 
-    async def _execute_tool_with_llm(self, variables: KernelArguments) -> str:
-        """Execute the appropriate tool based on LLM reasoning."""
+    async def _execute_with_azure_ai_agent(self, step: Step, action_request: ActionRequest) -> str:
+        """Execute the request using Azure AI Agent.
+        
+        Args:
+            step: The step to execute
+            action_request: The action request
+            
+        Returns:
+            The result content
+        """
+        # Create chat history for the agent
+        messages = []
+        
+        if step.human_feedback:
+            messages.append({
+                "role": "user", 
+                "content": f"Task: {action_request.action}\n\nHuman feedback: {step.human_feedback}"
+            })
+        else:
+            messages.append({
+                "role": "user", 
+                "content": f"Task: {action_request.action}\n\nPlease complete this task."
+            })
+        
+        # Pass context to the agent execution
+        execution_settings = {
+            "step_id": action_request.step_id,
+            "session_id": action_request.session_id,
+            "plan_id": action_request.plan_id,
+            "action": action_request.action,
+        }
+        
+        # Execute the agent with the messages
+        result = await self._agent.invoke_async(
+            messages=messages,
+            kernel_arguments=KernelArguments(**execution_settings)
+        )
+        
+        # Extract the result content
+        return result.value
+
+    async def _execute_with_function_calling(self, step: Step, action_request: ActionRequest) -> str:
+        """Execute the request using regular function calling.
+        
+        Args:
+            step: The step to execute
+            action_request: The action request
+            
+        Returns:
+            The result content
+        """
+        # Update chat history
+        self._chat_history.extend([
+            {"role": "assistant", "content": action_request.action, "name": "GroupChatManager"},
+            {"role": "user", "content": f"{step.human_feedback or 'Please complete this task'}. Now make the function call", "name": "HumanAgent"},
+        ])
+        
+        # Set up variables
+        variables = KernelArguments()
+        variables["step_id"] = action_request.step_id
+        variables["session_id"] = action_request.session_id
+        variables["plan_id"] = action_request.plan_id
+        variables["action"] = action_request.action
+        variables["chat_history"] = str(self._chat_history)
+        
+        # Execute with LLM planner
         planner = self._kernel.func("planner", "execute_with_tool")
         
         tool_descriptions = "\n".join([f"- {tool.name}: {tool.description}" for tool in self._tools])
@@ -319,7 +403,9 @@ class BaseAgent(KernelBaseModel):
             logging.warning(f"No connector service found for {group_chat_manager_id}")
 
     def save_state(self) -> Mapping[str, Any]:
+        """Save agent state for persistence."""
         return {"memory": self._memory_store.save_state()}
 
     def load_state(self, state: Mapping[str, Any]) -> None:
+        """Load agent state from persistence."""
         self._memory_store.load_state(state["memory"])
